extract tabulate functionality to separate module
