+++ src/main.js
@@ -1,4 +1,5 @@
 import { margin, width, height } from './modules/utilities.js';
+import { classify, benchmark, tabulate } from './modules/table_creator.js'
 
 // This step is performed to parse the url to identify the dataset and the default coloring column
 var query = window.location.search.substring(1);
@@ -126,63 +127,6 @@ var searchdic = function(arri, find) {
   }
   return true;
 }
-
-// displays the summary in tabular form
-var tabulate = function(data_tab, columns) {
-
-  var table = d3.select("body").append("table")
-  .attr("class", "select3"),
-  thead = table.append("thead"),
-  tbody = table.append("tbody");
-
-    // append the header row
-    thead.append("tr")
-    .selectAll("th")
-    .data(columns)
-    .enter()
-    .append("th")
-    .text(function(column) { return column; });
-
-    // create a row for each object in the data
-    var rows = tbody.selectAll("tr")
-    .data(data_tab)
-    .enter()
-    .append("tr");
-
-    // create a cell in each row for each column
-    var cells = rows.selectAll("td")
-    .data(function(row) {
-      return columns.map(function(column) {
-        return {column: column, value: row[column]};
-      });
-    })
-    .enter()
-    .append("td")
-        .attr("style", "font-family: Courier") // sets the font style
-        .html(function(d) { return d.value; });
-
-    /*
-      crossfilter dimensions and group by
-      http://animateddata.co.uk/articles/crossfilter/
-      */
-      var output = "";
-      var cf = crossfilter(data_tab);
-      /* crossfilter currently only supports up to 32 columns) */
-      for (var i=0;i<columns.length && i<32;i++) {
-        var byParty = cf.dimension(function(p) {
-          return p[columns[i]]; });
-        output = output + "<b>" +columns[i] + "</b>" + "<br>";
-        var groupByParty = byParty.group();
-        groupByParty.top(5).forEach(function(p, i) {
-          output = output + p.key + ": " + p.value + "<br>";
-          console.log(p.key + ": " + p.value);
-        });
-        output = output + "<br>";
-      }
-    // side table
-    document.getElementById("demo3").innerHTML = output;
-    return table;
-  }
 
 // create the dropdown menu
 // Coloring
@@ -565,8 +509,8 @@ function highlighting(val_search, val_transp, val_opacityMatch, val_opacityNoMat
 
         // get values for table -> array inside a list
         var zsx = lasso.items().filter(function(d) {return d.selected===true});
-        x_values = [];
-        y_values = [];
+        let x_values = [];
+        let y_values = [];
         // adjust the x and y values
         for (var i=0; i<zsx[0].length; i++) {
           x_values.push(((((zsx[0][i].getBBox().x+6.5) * (x_max - x_min))/width + x_min )));
@@ -582,7 +526,7 @@ function highlighting(val_search, val_transp, val_opacityMatch, val_opacityNoMat
             x_values[ii] = +(x_values[ii].toFixed(3));
             y_values[ii] = +(y_values[ii].toFixed(5));
             if ( (x_values[ii] === +(temp1[jj].toFixed(3))) && (y_values[ii] === +(temp2[jj].toFixed(5))) ) {
-              all_values = {};
+              let all_values = {};
               for (var k=1;k<categories.length;k++) {
                 all_values[categories[k]] = (dict1[categories[k]][jj]);
               }
@@ -633,55 +577,6 @@ function highlighting(val_search, val_transp, val_opacityMatch, val_opacityNoMat
 
     // Init the lasso object on the svg:g that contains the dots
     svg.call(lasso);
-
-    var classify = function(indices, vs_source, weights_source, biases_source, vocab_source) {
-      var sum_vectors = vs_source[indices[0]];
-      for (i=1; i < indices.length; i++) {
-        sum_vectors = math.add(sum_vectors, vs_source[indices[i]]);
-      }
-      var avg_vector = sum_vectors.map(function(x) { return x / indices.length; });
-      var mul = math.multiply(avg_vector, weights_source);
-      var add = math.add(mul, biases_source);
-        // get indices of 10 greatest elements
-        var topIndices = findIndicesOfMax(add, 10);
-        console.log("Top predicted vocab words:");
-        var strbuilder = "Predicted words:";
-        for (i=0; i<10; i++) {
-          console.log((i+1) + ": " + vocab_source[topIndices[i]]);
-          strbuilder += " " + vocab_source[topIndices[i]] + ",";
-        }
-        document.getElementById("predicted_words").innerHTML = strbuilder.slice(0, -1);
-      }
-    var benchmark = function(indices, bow_source, vocab_source) { // may need to adjust
-      var num_indices = indices.length;
-      var vocab_freq = bow_source[indices[0]];
-      for (i=1; i < num_indices; i++) {
-        vocab_freq = math.add(vocab_freq, bow_source[indices[i]]);
-      }
-        // get indices of 10 greatest elements
-        var topIndices = findIndicesOfMax(vocab_freq, 10);
-        console.log("Most frequent words:");
-        var strbuilder = "Most frequent words:";
-        for (i=0; i<10; i++) {
-          console.log((i+1) + ": " + vocab_source[topIndices[i]]);
-          strbuilder += " " + vocab_source[topIndices[i]] + ",";
-        }
-        document.getElementById("frequent_words").innerHTML = strbuilder.slice(0, -1);
-      }
-
-    // Utility function used for predicting words in semantic setting
-    // Source: https://stackoverflow.com/a/11792230/7100714
-    function findIndicesOfMax(inp, count) {
-      var outp = new Array();
-      for (var i = 0; i < inp.length; i++) {
-        outp.push(i);
-        if (outp.length > count) {
-          outp.sort(function(a, b) { return inp[b] - inp[a]; });
-          outp.pop();
-        }
-      }
-      return outp;
-    }
 
     console.log('Loading main data, again') // load data
     d3.tsv(dataset, function(error, data) {
